smother zoom movement

+++ src/rvlib/domain.rs
@@ -471,8 +471,8 @@ impl BB {
         shape: Shape,
         oob_mode: OutOfBoundsMode,
     ) -> Option<Self> {
-        let x_shift: i32 = to.x as i32 - from.x as i32;
-        let y_shift: i32 = to.y as i32 - from.y as i32;
+        let x_shift: i32 = (to.x  - from.x) as i32;
+        let y_shift: i32 = (to.y - from.y) as i32;
         self.translate(x_shift, y_shift, shape, oob_mode)
     }
